Switch to Freescale Release BSP 1.7 (Dizzy)
The Freescale Community BSP does not provide support for browsers and does
not have full hardware acceleration for 3D and multimedia, e.g. in Qt5.
Therefore switch to the real Freescale Release BSP that is supporting more
hardware features of the Freescale SoCs (i.MX6, etc.). Unfortunately this
release is still based on Yocto 1.7 (Dizzy).