Describe new way to configure build directories using DISTRO
Starting with Yocto 2.0, Freescale uses a new way of configuring the Yocto
system for a build.

  DISTRO=<distro> MACHINE=<board-name> . fsl-setup-release.sh -b <build-dir>

Replace the existing method in README-yocto-f+s.txt with the new one. Also
describe the possible distros and update the list of supported F&S boards.

Also generally update the readme for version Y0.4.