Rename meta-f+s to meta-fus
Having the plus character in meta-f+s may cause problems when using regular
expressions. To be safe, rename meta-f+s to meta-fus. This also matches the
fus we are appending to generic source code packages now. Change the file
yocto-download accordingly to add meta-fus to bblayers.conf.

Also change the description in yocto-download of how to configure for an F&S
environment by using distros and architecture family names now instead of
board names in the past.