Add fsgetty and audio chip configuration
- Add fsgetty from buildroot (fsimx6ul-V2.1). To support this feature we
have to add a new inittab file. This file is copied by the sysvinit-
inittab_2.88dsf.bbappend file. To be sure to copy our inittab file, we
must ensure that our BBFILE_PRIORITY is higher then the priority of
all other inittab...bbappend files e.g. meta-fsl-bsp-release/imx/meta-
bsp (priority 8). Furthermore we added the S01fssetup start script to
the different run levels rc1-rc5. S01fssetup is also taken from
buildrooot (fsimx6ul-V2.1).
- Add Audio Chip configuration support. In the file asound.state there
are the configurations for the sgtl5000 and wm9715. We have to copy
this file to the target folder that it is available in the buildroot.
At the end of the bootprocess the start script S01fssetup check if the
file asound.state exists and if the package alsactl exists. If both
exists we store the configuration of the asound.state file.