Add package silex-wlan-fs for F&S boards with SX/UL CPU
The Silex WLAN chip currently only has a closed source driver and also
needs specific versions of the WPA and hostapd tools. This recipe, adds
the driver and firmware files to the target. So the provided silex-wlan-
fs recipe will copy all the requested files to the target directory.

The Silex QCA9377 WLAN driver also needs an F&S MAC address because
there is no MAC address pre-programmed. Unfortunately the only way to
set the MAC address is by creating a file
/lib/firmware/wlan/wlan_mac.bin with an entry of the form

      Intf0MacAddress=aabbccddeeff

Of course this file has to be different on each board so it would have
to be created on-the-fly. But this is not possible on a read-only root
filesystem. However F&S U-Boot can dynamically generate an appropriate
file that will show up in Linux as /sys/bdinfo/Silex-MAC. So basically a
symbolic link pointing to this file would be sufficient. However the
firmware loading mechanism that is used by the driver to load this file
can not deal with sysfs directly because it can not determine the file
size.

So this recipe will set up /lib/firmware/wlan/wlan_mac.bin on the target
to be a symbolic link pointing to /run/Silex-MAC and a small start
script S02silex (in /etc/init.d/ on the target) will copy the file from
/sys/bdinfo to /run early in the boot process. /run is a RAM disk
(tmpfs), so this will also work on a read-only root filesystem.

The recipe does not contain the tarball, or in our case the pre-built
binaries. Yocto usually downloads the main package from the Internet. In
our case, the binary archive is part of the Yocto based F&S release in
the dl directory. It has to be moved to Yocto´s download directory,
which is usually called downloads. There Yocto will find and use it. If
the release is extracted with the yocto_setup.sh script, then the script
will do this automatically.

The tar.bz2 archive contains binaries of the pre-built kernel modules,
firmware and configuration files and also the binaries of the Silex
specific versions of the WPA and hostapd tools, i.e. wpa_supplicant,
wpa_cli, wpa_passphrase, hostapd and hostapd_cli.