Move README-yocto-f+s.txt to fsbuild
README-yocto-f+s.txt contains the release notes of the whole Yocto release.
Having this as part of one of the source code packages is not correct. Move
the file to the F&S Release Build System fsbuild that will store the
(individually updated) file in the top directory of each release.

This package here still has the file History-F+S-Yocto.txt which shows the
local progress/history.