Fix recipe failure optee-os-imx_git.bb
This package will be automatically installed. I tried to remove this but
it doesnt work. For now we modify this recipe via sed command to support
our machine files.

TODO:
This is not the final version just a quick and dirty version to compile
yocto. Search for a better way to remove this recipe.