Add read-only filesystem option to fus-images
Until now all fus-yocto images had read-writeable rootfile systems by
default. But the linux-kernel mounted the rootfs as read-only as it was
set in the FuS u-boot parameter "mode". This led to an ubifs assertion
when booting the rootfs for the first time and the u-boot parameter was
ignored.

Now the user has to decide at build time, whether he wants an rw or ro
rootfs. This can be done by handing over the option "-m ro" or "-m rw"
to the fus-setup-release.sh script. The  default will be ro.
Also in an existing built the mode can be changed by adding or removing
the line
EXTRA_IMAGE_FEATURES += "read-only-rootfs" from the local.conf

The default "mode"-variable in the uboot will be set automatically and
should not be changed afterward. The roofs can still be remounted as rw
or ro after the boot process.