Add fus-image-mono
The fus mono image should help customers to check whether their
.NET-framework application is portable to Linux or not.
It has the open-source framework Mono preinstalled so the user can try
to run their .NET application and see if all the necessary libraries are
available.

It is only available for x11 for now because of the dependency of
libgdiplus.