Improve naming of Linux-kernel and Uboot in recipes
The linux kernel and uboot now have a generic name in the recipes.
This simplifies the automatic build process because of the names of the
final linux/uboot versions are no longer required in advance.

Now a softlink to respective verison has to be be created in the release
directory:

linux-fus.tar.bz2 -> Linux-kerenel
u-boot-fus.tar.bz2 -> Uboot
u-boot-fsimx7ulp.tar.bz2 -> Uboot-7ulp

The name of the directory inside the archive has to be:
linux-fus         for the Linux-kerenel
u-boot-fus        for the Uboot
u-boot-fsimx7ulp  for the Uboot-7ulp