Fix failed hostname service on armstonea9r2
We disable the automatic start of the bluetooth.service, because
it starts the bluetoothd daemon, which at this stage tries to
get the host name via hostnamed.service. The net namespace feature
is not activated in the Linux Kernel by now, so it fails.
The bluetoothd daemon has to be stated by hand after boot up.