Improve access to firmware marvel sd8997
We had an own recipe file called linux-firmware-wlan which copies the
firmware sdsd8997_combo_v4.bin from the linux-firmware package to
rootfs. In the past this firmware was not integrated because the
linux-firmware package was to old. Therefore we integrated an own recipe
"linux-firmware-wlan". In the meantime there is a newer version of
linux-firmware so we don´t have to use an own recipe. Copy the firmware
sdsd8997_combo_v4.bin from linux-firmware-%.bbappend file to rootfs
instead of using an own recipe.