Improve FS FUS_LOGIN_CONSOLE setup
In poky layer the variable SERIAL_CONSOLE is deprecated and removed. The
new variable is named SERIAL_CONSOLES. This is necessary that the login
console can be dynamically set by our bdinfo struct. This holds the
current login console for board. After switching to SERIAL_CONSOLES we
have the following problem. Now you cant set the login_tty and
serial_tty to different uarts. It must be same or you are unable to
login.