Improve default network startup for F&S images
By default yocto activate the network interfaces while booting. We also
have a ethernet switch which creates VLANs on some of the boards. But by
default the network will only be activated for the following interface
names "eth*" and "en*". The VLANs are called "lan*". So add own network
script which also activates "lan*" if they´re available.