Improve KERNEL_MODULE_PACKAGE_SUFFIX for layer meta-fus
This is necessary because there will be some debian packages created and
these packages must be lower case. Basically we set CONFIG_LOCALVERSION
to "-F+S". This appends the string "-F+S". Due to the string is upper
case we need to set variable to convert it to lower case.