Fix passing of ethaddr for Silex WLAN module
For Yocto the script S02silex now is replaced by the systemd unit
silex.service, which can already be enabled through a postinstall script
in the recipe.

The module loads the ethaddr through /lib/firmware/wlan/wlan_mac.bin, so
we need a softlink to the RAM area or else it is not possible to change
the ethaddr in a read-only filesystem.