Add error handling for the install script with eMMC
Currently the UBoot update process does not react to errors in UBoot
commands issued in the install and update scripts. So whenever the load
command fails, the RAM area will get written to the destination as it
is. For boards with a NAND flash this is not as problematic, because the
UBoot will not be affected by the update. For eMMC however the UBoot is
included in the sysimg file. An error in the load command results in a
corrupt UBoot and the board will not be bootable without external
flashing software.

The updated install script checks for errors that might occur in the
load command and aborts the update process before actually writing the
corrupt data. Also the chunksize for the eMMC update process is reduced
to support more USB sticks.