Add M4 image to the emmc sysimg file of fsimx7ulp
The default uboot configuration looks for the M4 image in the boot
partition of the emmc sysimg. Formerly, the M4 image had to be added by
hand.
Now M4 image from the release binary directory gets automatically added
to the emmmc sysimg.

Also the NXP M4 example images are removed from the build, because it is
not guaranteed that they will work with out boards.