Improve machine configs
We used to set a lot of machine settings in the distro files because the
values would get overwritten by the nxp layers otherwise.

Now we can set the settings in our machine files, which is more
convenient.

For this we introduce a ne include file fus-base.inc which contains
common setting like the kernel and uboot provider.

This file gets included in all the other machine files.

The fus-common-inc file only contains distro specific settings.

Also, update the include pathes of all machine files.