Improve linux-fus recipe
We want the linux version to be independent of the meta-fus layer to be
able to create yocto releases with kernel adaptions without having to
change the meta-fus layer.

To achieve this, the setup-yocto script will copy a linux-fus.bbappend
file to the meta-fus layer, which contains the SRC_URI and the SRCREV of
the kernel.

Also adapt the recipe to the new yocto version and remove the old
linux-fus.bbappend.

Also the package sd8997 is now part of the linux-firmware by default and
does not need to be included in the bbappend anymore