Add recipe for qtwayland on the meta-fus layer
When qtwayland is build, it downloads the qtwayland repository from
git://code.qt.io/qt/qtwayland.git and checks out
99efdbe19eab1811f22fc2ed12ff13f664633ba7. There is a patch missing that
does not allow it to compile because it will not find libdrm.

For this reason add a meta-fus appandage for the recipe that changes the
version that is checked out to the most recent at the time of testing
since the problem has been fixed in the current release with id
f2c225598c98f3f3e489fdcbd09c61e14d540f9b.