Improve Alsa Configuration File
The current configuration File will not be loaded, presumably because of
a change of the Software. This causes default Values to load that Mute
the Sound Output on Line-Out. It can be reactivated again by Hand and
saved to the System as a new default, but it should be unmuted by
default.

Add a newer configuration File that can be loaded in the current
version, and is unmuted on Line-Out by default.