Improve uboot-fus recipe
We want the uboot version to be independent of the meta-fus layer to be
able to create yocto releases with uboot adaptions without having to
change the meta-fus layer.

To achieve this, the setup-yocto script will copy a uboot-fus.bbappend
file to the meta-fus layer, which contains the SRC_URI and the SRCREV of
the uboot.

Also adapt the recipe to the new yocto version.